ticker working
