+++ client/src/containers/StocksInfoContainer.js
@@ -4,6 +4,7 @@ import Highcharts from 'highcharts/highstock'
 import HighchartsReact from 'highcharts-react-official'
 import StockTicker from "../components/StockTicker"
 import StockChart from "../StocksInfoComponents/StockChart"
+import { getDailyTickerDataBySymbol } from "../services/TickerService"
 
 const StocksInfoContainer = () => {
 
@@ -53,20 +54,66 @@ const StocksInfoContainer = () => {
                 setMETADaily(allDaily)
             })
 
+            const copyTicker = [...ticker]
+
+            getDailyTickerDataBySymbol('IBM')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'IBM')
+                    copyTicker.push(tickerObject)
+                })
+            getDailyTickerDataBySymbol('AAPL')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'AAPL')
+                    copyTicker.push(tickerObject)
+                })
+            getDailyTickerDataBySymbol('MSFT')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'MSFT')
+                    copyTicker.push(tickerObject)            
+                })
+            getDailyTickerDataBySymbol('TSLA')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'TSLA')
+                    copyTicker.push(tickerObject)            
+                })
+            getDailyTickerDataBySymbol('XOM')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'XOM')
+                    copyTicker.push(tickerObject)            
+                })
+            getDailyTickerDataBySymbol('WMT')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'WMT')
+                    copyTicker.push(tickerObject)
+                })
+            getDailyTickerDataBySymbol('NVDA')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'NVDA')
+                    copyTicker.push(tickerObject)              
+                })
+            getDailyTickerDataBySymbol('META')
+                .then(tickerData => {
+                    const tickerObject = createTickerData(tickerData, 'META')
+                    copyTicker.push(tickerObject)
+                })
+    
+            setTicker(copyTicker)
+
     }, [])
 
-    // const createTickerData = (loadedData, label) => {
-    //     let tickerDays = loadedData.slice(-2)
-    //     let todaysValue = tickerDays[1].y
-    //     let yesterdaysValue = tickerDays[0].y
-    //     let differenceFromDayBefore = (todaysValue - yesterdaysValue)
-    //     let differencePercentage = (differenceFromDayBefore / yesterdaysValue) * 100
-    //     const tickerObject = {
-    //         name: label,
-    //         value: todaysValue.toFixed(2),
-    //         percentage: differencePercentage.toFixed(2)
-    //     }
-    //     return tickerObject
+    const createTickerData = (loadedData, label) => {
+        let todaysValue = loadedData[0].value
+        let yesterdaysValue = loadedData[1].value
+        let differenceFromDayBefore = (todaysValue - yesterdaysValue)
+        let differencePercentage = (differenceFromDayBefore / yesterdaysValue) * 100
+        const tickerObject = {
+            name: label,
+            value: todaysValue.toFixed(2),
+            percentage: differencePercentage.toFixed(2)
+        }
+        return tickerObject
+
+    }
 
     
    return(
